Hi.

I have rewrite a little scripts (in '**sources**' dir) and got speed boost more than on 50%.
I think it will improve your building timings in `CI/CD`.

Results on `Intel(R) Core(TM) i3-8100 CPU @ 3.60GHz` (4 cores):

| Script | Execution time BEFORE | Execution time AFTER | Speed boost |
| ------------- | ------------- | ------------- | ------------- |
| sources/`build-statics.sh`  | 2.54 min. |  1.10 min. | faster more than on **`59%`** |
| sources/`build-vf.sh`  | 24 sec.  | 13 sec. | faster more than on **`45%`** |
| sources/`build-webfonts.sh`  | 21 sec.  | 6 sec. | faster more than on **`70%`** |

Proofs:

```bash
###########################################################################
# Timings BEFORE:

manlix@lab:~/git/mono$ time ./sources/build-statics.sh
...
real	2m54.585s     <---
user	3m29.372s
sys	0m4.632s

# Timings AFTER:

manlix@lab:~/git/mono$ time ./sources/build-statics.sh
...
real	1m10.804s     <-------------------------- faster more than on 59%
user	3m46.461s
sys	0m5.748s

###########################################################################
# Timings BEFORE:

manlix@lab:~/git/mono$ time ./sources/build-vf.sh
...
real	0m24.533s     <---
user	0m23.745s
sys	0m0.777s


# Timings AFTER:

manlix@lab:~/git/mono$ time ./sources/build-vf.sh
...
real	0m13.012s     <-------------------------- faster more than on 45%
user	0m24.797s
sys	0m0.845s

###########################################################################
# Timings BEFORE:

manlix@lab:~/git/mono$ time ./sources/build-webfonts.sh
...
real	0m21.728s     <---
user	0m21.275s
sys	0m0.424s


# Timings AFTER:

manlix@lab:~/git/mono$ time ./sources/build-webfonts.sh
...
real	0m6.131s     <-------------------------- faster more than on 70%
user	0m22.977s
sys	0m0.465s

###########################################################################
```

_p.s. Tested on Ubuntu 20.04.2 LTS_