@philippnurullin 2021-09-14
The build scripts are changed drastically so this PR is not relevant any more.
Sorry, closing it.